Started development
I started the development of the light version of the eBits B2B sales system. I initialized a React.js app and I started installing the dependencies needed. I added Tailwind CSS to the project, I added the react-router-dom library. I started writing the code by initializing the routes of the application. I wrote the code for the context components as well. I started developing some of the UI components as well