Two components to help dealing with bugs and scalability
Developed the kit handler component. The kit handler component helps reduce boilerplate code, as well it makes it easier handle possible bugs that would show when changing the state of the receipt from products to a kit. I as well designed and developed a amount handler component. It is being used instead of the two buttons for reducing and increasing the product amount. It reduces the boilerplate code and helps defining and dealing with possible bugs. It also helps maintain reusability. I as well fixed some minor design bugs.