The name kit page
I made the name kit page, which appears only if the user selects they want to export the query as a custom link. I as well developed the page left side of the name it page, and the component that represents the receipt. The receipt is a static component, it only displays the products and the price, the user cannot interact with it. I developed the atomic component which will display each of the products names and prices.