Styling changes after testing
I did some styling, I fixed some present styling bugs, as well as implemented some of the new ideas gained after testing. I fixed the sizing of the item in the basket on page left side in the product selection page. I fixed the positioning of the magnifying glass icon on the page left side and product search bar. I fixed the misalignment of the two columns in page left side static on the contacts page (extended). I added a grey background with low opacity behind the pop-up message on final check page. I fixed the styling bug with the overflowing of search result items on the pop-up message as well.