First fix after testing
After the testing, I started fixing bugs and implementing the insight gained during the tests.  I fixed the alignment of the text in the info pop-up message. I designed and developed a component, that handles direct user input on the amount handler. Now the user is able to not just spam plus and minus icons, but also to simply just type in the amount they need to order.  To do that I needed to add a new function in the ShoppingCartFunc context component. It is called "handleDirectAmountInput" it takes two arguments, product and amount. Also I fixed the problem with the missing pictures of some of the product variations. I developed an atomic component called ProductImageHandler. To it props are being passed, the variation image url and the base product image url. The base product image url is a fallbackURL. The image onError dom property triggers a function, that swaps the two images with the help of a state variable that manages them.
I fixed the product display section on the receipt to be full height. The height is handled dynamically. If the query is for a kit, it reduces the height so it can display the bottom section of the receipt.