Small adjustments
Fixed some small bugs and re-factored the code a little. Removed unnecessary code and changed the code to be even more dynamic and easy to maintain.  For example, I made the buttonsholder component accept title and navigation destination adjustments, only optionally. I as well fixed some small styling mistakes and made the page look more welcoming and interactive. Added some animista animations. Fixed the static receipt of the name-it page as well. Made the table on the last check page to be scrollable and fixed the small styling bugs there.